requirements.txt
torch>=1.9.0
transformers>=4.20.0
peft>=0.3.0
datasets>=2.0.0
flask>=2.0.0
flask-cors>=3.0.0
requests>=2.25.0
beautifulsoup4>=4.9.0
pdfplumber>=0.6.0
rouge-score>=0.1.0
nltk>=3.6.0
psutil>=5.8.0
pyyaml>=5.4.0
accelerate>=0.20.0
bitsandbytes>=0.39.0
scipy>=1.7.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.62.0
wandb>=0.12.0
tensorboard>=2.8.0

